pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pet-ota"
version = "1.0.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "bsdiff4>=1.2.0,<2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tenacity",
    "structlog",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]
signing = [
    "pet-quantize>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["pet_ota", "pet_ota.*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[[tool.mypy.overrides]]
module = ["bsdiff4", "pet_quantize.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
